Force no animation when setting the viewport
Our usage of `set_viewport_position` always is to fully control the
viewport, often to prevent any scrolling.  I thought the default
here is `False` but actually the default is to follow the global
`animation_enabled` setting (which in turn is `true` by default).

We hence force `False` in all calls.

This removes some flicker and glitches when the user has animation on.