Fix reverse tab cycling
Since b9e9928 (Rewrite settings abstraction for performance and
reliability) settings are cached and MUST NOT be mutated.

The code base is small here so instead of sending a copy of a setting
on access it is sufficient to make this small change here. We already
do not mutate settings anywhere else.