Modernize syntax settings of our special views
The general move here is to turn the "gutter" on because otherwise
Sublime will never highlight the current line.  But because we don't
show line numbers in our dashboards we shift the view content to the
left (with a negative margin) to reduce the empty whitespace area.

For some dashboards it looks nice to have the fold buttons on.  This
really just adds some visual guidance, although you can fold out for
example the stashes section using these.  Note that we must turn off
"fade_fold_buttons", because the negative margin shifts the hover area
of the buttons away from them.