Show the whole hunk when using `gs_next_hunk`/`gs_prev_hunk`
We used to just show the first line of the hunk which is a mistake.
(Iirc this was just done to do less work; a kind of performance fear
without ever measuring.)

We now walk and collect the whole hunk and compute a combined region
for it.

Note that `gs_prev_hunk` already walked the whole hunk before; without
a noticeable performance issue.