Fix: Allow wide jumps if on the next to last commit of a chain
If the user is just on the sibling commit before a commit chain ends,
we didn't move to the end of the chain.

In that case `next_dots` is already exhausted (because there is only
exactly one next dot which we already pulled).

We fix by renaming all local variables for clarity.  After that using
`next_next_dot` in the `else` clause will throw because the for loop
did never run.  We catch and do nothing because `next_dot` will now
already contain what we want, the next dot.