Adjust scrolling after wide jumps in the log graph
After a26f976 (Improve `show_region`) wide jumps could easily lead to
overscrolling because in the log graph we only passed the target line
to highlight.  (Then, if that target is out of the visible region,
`show_region` would try to show it around center position which is not
what we want.)

Conceptually, we want to show the original dot, the position we jump
*from*, and the target dot we jump *to*.  For that, we compute the
combined region of both using `join_regions`.

Unfortunately, `show_region` always preferred the beginning of the
region. We thus teach it to either prefer one or the other side of the
passed in region.