Do not cache reading the HEAD commit
Ref #1368

In other places, we check if the `commit_hash` is `""` or `"HEAD"`
to skip caching, so this one was missing here.

The use case is a user binding to `gs_show_commit` with `commit_hash`
set to `HEAD`.  This kind-of just works with this patch.

However, arbitrary references, like for example tags or branch names,
are not supported.