Let `[F]` toggle all filters
`F` used to reset only the verbatim "filters" section.  Let it
instead toggle "paths", "filter_by_author", and "filters".

For this, we introduce new state `apply_filters` and use it when
rendering the content.  After that the `gs_log_graph_reset_filters`
implementation gets a trivial switch.

Note that we modify the behavior when editing "filters" or "paths".
Whenever "apply_filters" was `False` *before* running, we actually
*reset* all other filters because that's what the user saw.