Rework when to reuse graph views
Since we got more dynamic now we want to reuse open graph views more
aggressively.

For example, 1. if a graph view does not apply possible filters, do
*not* consider these "hidden" filters when comparing.

For example, 2. if a standard (`all: True`) graph is open, use it for
"File History".  The user might `F` on such a view, after that treat
the view again as our standard graph.  Reusing it on the typical `g`,
or for a different "File History" etc.

This is likely a WIP, ideally we want to reuse as much views as
possible, maybe only allowing one GRAPH at any time.