Force focus for native git files on load
If "subl -w" is set, Sublime has problems focusing the active view
but only if we start the git command from within Sublime itself, for
example using GitSavvy itself or Terminus.

We mitigate by trying to force focus the just loaded view.  (A way to
bring the view's window to front would be handy here.)