Fix possible freeze when undoing in the commit message view
The interaction between automatically setting the `read_only` state and
automatically updating the diff in the background confuses Sublime
in a way that it freezes and shuts down.

A reproduction was:  Put the cursor in the read only area, for example
the help comment.  Try `ctrl+z` for undo while in the read-only area.
Note that it tells you in the status bar "Undo: ..." without any visual
difference.  Move to the writeable area, `ctrl+z` again.  Now, Sublime
freezes.

The fix is to switch the `read-only` state not sync, t.i. using the
`on_selection_modified_async` callback.  From the perspective of the
fix it looks like "undoing" triggers a cursor movement *to* the
read-only area, and a sync selection-modified event then toggles the
read-only state which crashes Sublime.