Implement generic "gs_commit_log_helper"
Fixes #487

Replace `gs_fixup_helper` with a more generic `gs_commit_log_helper`.
Add a binding at BOF position to `<tab>` which will just open the log
panel to quickly see what has been done.  As before, `enter` will copy
the commit subject for reuse.

The "fixup" and "squash" helpers are now special cases.

Note that for the latter the cursor ends on EOL position. For the
former it stays on BOL, and a next `tab` will again open the helper.