Fix "make_commit" syntax tests
The message body and subject line tests had no test prefix, so they
didn't actually run.  Note that a true first line test is not possible,
because the testing framework applies the syntax to the first comment
line already.  (T.i. the definition "# SYNTAX TEST..." is already the
subject line.)  Other than use `#` instead of `//` as a test marker for
consistency.