For some commands the input should be the non-inclusive start commit
In the first iteration we always applied "^" to the given commit.
However, it makes more sense for the user to run some commands with
the typical start commit which is non-inclusive in git.  For example,
`master...HEAD` does *not* include the master commit.

For some commands, like "edit/reword/drop X" we expect the commit to be
edited, reworded etc. as input.  (And then we internally append "^"
because that's an implementation detail of the how "we" or the rebasing
feature works.)  On the other side, for e.g. `rebase X --onto Y` we now
expect a starting commit (from X on).  For example autosquashing
`master` is for the end-user more convenient to use because the first
commit after master is even hard to compute.

Note that this does not affect usage of the menu and bindings.  We only
make the usage of these commands simpler.