Do not reuse the standard graph view if `branches` get selected
Although switching from a standard "all" graph to a graph which
compares branches is what we want in the long run.  Right now, there
is no easy way to undo that.  Esp. `[b]` is still alpha and not
documented.

To undo, a user would have to invoke `[a]` and then `[b]`, delete
everything and commit with `<enter>`.  This is too much so we prohibit
this case of reusing and open a new graph view instead.