Fine tune fixup and squash helper triggers
1. Allow the helpers at EOL position if the line is a fixup or squash
line.

This is actually the position the user lands on after using the helper,
and we want to allow selecting a different commit easily.

2. Only trigger the helper if there no text follows.

Here, for example the user first wanted to do a "fixup" but then decides
to edit the first word from "fixup" to "squash".  The user therefore
goes to BOL, deletes the right word, and types "sq<tab>" to autocomplete
to "squash".  This just works now because "squash" is via the help text
always in the buffer.