Fix: Compare longest with shortest commit message
We currently work visually here, t.i. on the actual drawn graph, and
have to expect shortened commit messages.  We always worked with a
cleaned original message (see: `message_from_fixup_squash_line`) but
have to do the same (`.rstrip(".")`) when comparing with other messages.

The code also assumed the original message is the truncated one.
This is often but not necessary always true. We therefore sort by length
before comparing with `startswith`.