Fix: Graph menu should open if cursor is on a detached HEAD
Since 0bcabc7e (Add "fetch"/"push"/"pull" to the graph menu) published
with 2.26.0 (Sep 2020) a user cannot open the actions menu if on a sole
detached HEAD commit.

Generally, the idea of having a sparse dict to represent the line or
commit info makes the handling here brittle and hard to read.  We can
handle that later, for now `info["local_branches"]` can fail and must
be made safe using `.get()`.