Use a different storage for the "show_option_all" remote panel
With "show_option_all" the panel gets an imaginary first item, namely
"All...".  We want to make it possible to save this selection, and to
not interfere with all other remote panels which do not set this option.
(Currently, only `gs_fetch` uses it.)

For example, we want to support a user always fetching all but pushing
to (of course) a specific remote.

Enhances #791. We change the initial value back to the common "origin"
in all cases.  Generally, fetching all does not scale, and it is an
edge-case feature.