Assume `commit -a` if nothing is staged yet
As quality-of-life hack: If you enter the commit process with nothing
staged, assume `include_unstaged`.

This is transparent as we always commit what's visible in the diff in
the commit message view.  If `-a` is not what you want, the user can
either just close the view, stage something and usually using `[c]` in
some of the views enter the commit process again.  She can also just
switch to a diff view, stage something and hit `[c]`, t.i. without
closing the view, as the view will update.