Implement `select_all` helper for the commit view
We want to support especially `select_all` followed by `delete` because
that makes the commit message empty so that the user can safely abort
the commit process.

Since the commit view is a hybrid writable/read-only view this doesn't
work because `select_all` selects the whole buffer leaving the cursor
at EOF position which in turn switches the view to read-only mode.

We thus implement a helper, intercept the `select_all` calls, and just
select the commit message.  We're cautious though, and only do this if
nothing is selected and the cursor is within the writable message area.