Probe previous tip and scroll to it
Probe `@{1}` *before* updating the view state as otherwise "refresh"
will throw in an unpleasant way.  Especially it will not even redraw
the prelude/help section, and of course throw on each focus, so closing
the view would be the only option for the user.

Since we now have the sha just `follow` it which really helps with
orientation.