Try to find the exact line the cursor was on
Read the real position the cursor was on from the diff as a pair of
line numbers `(a_line, b_line)`.  Then, after refreshing, read in the
diff again and search for the same line id.

To identify the correct file, instead of extracting the filename on i.e.
the b-side, just take the whole first line of a header as a stable
identifier.  The first line does not change when "zooming" and we don't
have to deal with "/dev/null" pseudo-names which `head.from_filename()`
rejects.

Also capture the current scroll position, and apply it to keep the
viewport stable.