Create a fake line if the user is past the diff
The user can be at EOF position which is never *on a hunk* because
we always add a new empty line after the diff.  Support that by
constructing a "fake" line, imaginary just after the last line of the
hunk.

Assumes the loop was never empty, t.i. `line` is not unbound which
should be true because we first search for the `hunk_for _pt` and a
hunk can't be empty.