Refactor `HunkLineWithLineNo` type
Instead of a flat tuple `HunkLineWithLineNo` implement the more
interesting part of it, `LineId`, as a `namedtuple`.

Only capture and then pass around these `line_id`s for example in
`cur_hunks`, `find_line_in_diff`, as the `line` itself is only of
interest when comparing its `region` with the current cursor. (For
example we don't read its `.text` content anymore.)