Implement to stage, unstage, or discard concrete selections
Teach the diff view to stage, unstage, or discard selected regions.
This includes single lines and also multiple selections.

The strategy here is to reduce selections to zerodiff patches.  This
is grounded on a correct implementation for `recount_lines`, though
staying compatible with the previous `compute_line_ids_for_hunk`.

After forming good "U0" patches we can use the functions in `stage_hunk`
for further processing.

The difference when rewriting a series of hunks to make a valid patch
for reset or for staging is basically which start line, the one from
pre ("a") or post ("b") image, is the "main" line on which we compute
the opposite line number.

We use the same bindings `sud` and decide the strategy to use if there
is a selection made, or if there are only cursors currently.