Provide the empty string as default for `stdout` and `stderr`
If the git output is not decodable we throw a `GitSavvyError` without
providing `stdout` nor `stderr`.  This results in throwing a `TypeError`
in `commit.py` where we serach for a string  in `stderr` without
checking if `stderr` is not `None`.

Instead of doing the `None`-check in `commit.py` provide the empty
string as default here to avoid possible other checks throughout our
code base.

Ref #1426