Fix: Let the diff view catch decode errors again
Regression since 4060ae79 (Other: improve catching errors, Jul 2017)

The diff view catched a `GitSavvyError` and looked for a
`UnicodeDecodeError` in its first argument.  This is never the case
since the refactoring in 4060ae79.

Opt-out of automatic decoding using `decode=False`.  Instead of showing
a modal dialog in case of an error, just print a message in place of
the usual diff.