Do not open an inline diff if the diff is not decodable
In case of an `UnicodeDecodeError` we showed the modal dialog but kept
an empty view afterwards.

Rather catch the error and close the view automatically.  (We also
close the view automatically if the diff is a combined diff, t.i. the
file has a merge conflict currently.  So this behavior is consistent.)