Catch decode errors in the commit message view
Fixes #1426

Showing the diff is optional in the commit message view.  It is good to
have but a user can commit anyway.

In the case the diff is not decodable, do not show the error modal
dialog anymore.  Instead print an error message in place of the diff
just like we do in the diff view.