Fix live logging feature
Fixes #1437

For the live logging feature we basically implement `communicate` on our
own.  `communicate` generally works with three threads, two *reader*
threads, one for stdout, one for stderr, and one fan-in or *join*
thread.

Let's make the reader threads short reading loops over the buffer io
streams.  Zip both streams `out` and `err` to keep the order of output
close to the original order.  Use no-op classes `Out(bytes)` and
`Err(bytes)` to tag the output messages for later processing.

The fan-in thread is the thread which initiated the git call, it is in
a busy loop.  Log while unzipping the input lines.

Since the fan-in thread is already busy looping we cannot deadlock with
`M`.  Note that we basically do *not* log *live* if the initiating
thread is actually `M` because we cannot *force* Sublime to draw.
In that case, Sublime will wait for the main task to be finished and
draw once.

Remove the now superfluous `live_panel_output_timeout` setting.