Fix: Ensure rewrite actions run on the worker thread
Fixes #1418

Fix a regression introduced by 1e31e247 (Let `LogPanel` continue
synchronous on `enter`).

Commit 1e31e247 introduced that `show_log_panel` continues on the main
thread on `enter`.  Hence actual users of `show_log_panel` must
explicitly opt-in and hop to a different thread if they want to.

Even for the rebase dashboard running on the main thread had the
advantage that the user could not issue different commands *while*
a rewrite was in progress.  Generally though, these rewrites possible
take a long time so it can feel like Sublime freezes.  Running on the
worker might be the better solution here.

The actual bug #1418 describes is a bug in the long distance.  When the
user uses the "live_panel" panel feature for "git commit" commands,
we deadlock for "live_panel_output_timeout" (default: 10s) because the
reader thread used for live logging waits for the main thread to write
to the panel while the git commands - running on the main thread - waits
for the panel to update.

This commit does *not* solve this underlying bug-by-design but just
the problem at hand. This is still the right commit as rewriting on the
main thread can take too long time anyway.