Remove internal `show_status_message_on_stderr` arg
This setting is unused in the code base; it is only set to `False` if
we *also* set `show_panel_on_stderr` to `False`, basically to suppress
our automatic error handling.

Since popping up the output panel is already a big visual indicator,
it is even unlikely the user also looks and reads the status message
which disappears after 2 or 3 seconds.  The git output should also
clearly indicate that something and especially what went wrong so this
status message seems superfluous.