Remove `close_panel_for` setting
The intention here was to indicate that we start a new command.  Poorly
implemented, it just closed any panel at the start of a git command,
t.i. it did not check if the active panel actually is our output panel.

Since then we have the live logging panel which gets cleared at the
start of a command.  So we clearly indicate now when a command starts.

If the panel is actually an error panel, we'll find a better way to
close such panels automatically later.