Remove `run_async` argument when logging to the panel
The argument is almost not used and the feature seems fishy as writing
to the panel requires a lock on the main thread anyway.

Fixes a bug where for example for live logging "commit" the panel never
logged "[Done ...]" because after delaying this write the original view
is already gone.  Or, when rewriting using rebase, multiple "[Done ...]"
messages were written after the last commit.