Remove `show_stdin_in_output` setting
Typical usages of the live panel are "push", "pull", "rebase", "commit".
Only for "commit" we have some stdin but why should we duplicate it
here?  In the original PR the usage was on "apply".  But live logging
for "apply" is a non-starter.  It might be interesting while hacking on
our staging commands but that's not for end-users.

Remove the setting which was off by default anyway.