Do not cache the file name of ordinary views
I could not find a reason in the repo why we do this.  Obviously, if we
cache the filename in the view settings we don't "see" renames
immediately (but only after repopening a new view into the file).  This
for example leads to diffs against the old filename where naturally
every line is red (as deleted) and can be staged.  This feature seems
arbitrary as we can't stage any green additions.

As it's not clear which feature we might support here, remove it so we
naturally "follow" renames as expected.