Show progress panel on clone
Since "clone" is a rather long running process show the panel as a
progress indicator.  For that to look nice, first open a new empty
window which will host the new clone and the progress output panel.

Also fix the `refresh_gitsavvy_interfaces` invocation to run on the
new window not the original.