Reuse existing window if it's empty
As usually we pick a sibling folder of the current project as the new
projects root; in a different workflow let the user first open a new
window (for example `ctrl+shift+n`), then do `gs_clone`.  As there are
no folders open yet, we will suggest `~` as the root which is good for
temporary projects.

As long as the window is empty just reuse it.