Test if the reported filename is accessible
As an edge case, for example when viewing a file in a zip container,
Sublime Text might return a "correct" looking file path although the
later `resolve_path` will throw as the file does not exist.

That case was generally handled by a later call to `os.isdir`.  It is
only within this subroutine we might use the invalid filename.

Bug introduced by a5561c59 (Fallback to first folder only if necessary).