Fix extracting the project name from a remote url
The remote url can contain `.`s, and our algorithm did not expect
that.  It might not need to end with `.git` as well.  And it doesn't
have to be funky implemented.

Note that the old implementation indicated it would return the empty
string for pathological cases but this was never the case; the else
clause in the return statement was unreachable.