Let `gs_log_graph_action` use `gs_checkout_branch`
We want the log graph action to participate in the new "--merge"
feature, but also it should remember the "last used branch".

We didn't do this in the first place because `commit_hash` here can
be any commitish and maybe will lead to a detached HEAD.  In that case
we "pollute" the "last-used-branch" cache.  This is good enough though
as the cache can always have invalid entries, for example if we delete
a branch.