Place diff views next to the current view
Diff views are usually created or switched-to from the status dashboard
or any possible source file.  They can be reused in which case we just
switch to them.

If we create a new diff view it always is created on the right of the
current view, so that upon closing the diff view the user lands again on
the original view.

We want the same when switching (or "re-using") existing views so we
move the diff view in the correct position so that closing it works as
expected.