Remove unreachable statement
The reset command wizards first ask for a commit or branch, then for
the mode.  If the user escaped the mode question they used to go back
(or pop the stack) to the first question.

This has been removed in 217d2c06 (Add "reset to" action), probably as
WIP commit.  We can't just reintroduce it because when "commit_hash" is
given as an arg, t.i. what the "reset_to" functionality does, there is
actually no first step and we must abort the wizard.