Use `use_reset_mode` as a default mode
Fixes #1472

`use_reset_mode` was used to bypass the reset mode selection completely.
This is not very useful as a mode is not a user preference but depends
on what the user wants to do.

To complicate matters `use_reset_mode` was also used to *define* an
arbitrary new mode, e.g. "--mixed -N".  Because of that just removing
the setting was not an option right now.

To improve the situation: allow defining a new mode; but whatever is set
treat it as the preselected default mode.