Check `returncode` to decide if `git log` failed
Checking `returncode` and not if `stderr` is non-empty is consistent
with `git_command`'s behavior.

The previous behavior leads to a bug in conjunction with git v2.33 which
emits encoding warnings on stderr. Without this patch, we would throw
and fail although we actually lax decode git's output just fine. (This
behavior of git has been reverted in v2.34.)