Allow deletion of the active branch
As deleting the active branch is forbidden by git, offer to detach
first.

Typical usecase is that the user is on a feature branch which has been
merged upstream.  After `fetch`, the users sees the merge and can just
delete the feature branch without checking out a different branch
before.