Maybe update commit subject if reusing commit message view
When the user used a fixup helper and there was a commit message view
somewhere open, we correctly reused the view but did not add the
`initial_text` at all.

Let's fix this.  Do not completely replace a possible commit message
the user might have typed in already.  Just add the `initial_text`
at the front.