Move `last_branches` storage to the global store
Do not manage `last_branches` as a by product of `gs_checkout_branch`
but rather as a side-effect of `get_status`.  This has the benefit of
not missing an update in case the user checks out branches using
a different method -- for example using the command line.  We can also
use `gs_checkout_branch` now with commit hashes (not just branch names)
and be sure these do not land in the branches history.

Downside: if `get_status` is too slow, we loose the functionality more
or less as the history gets not updated frequently and fast enough.
Honestly, GitSavvy (and git) is often useless for such large repos
anyway.