Always checkout on the ui thread
The actual `checkout` action did run either on the worker or UI thread
depending on whether the command `gs_checkout_branch` was called with
a branch name - in that case it did run on the worker - or with `None`
- in which case we first show a quick panel and then continue on the
main UI thread via the `done` callback.

Let's run on the UI thread - blocking - always.  This might be changed
at a later time but for now blocking means the views do not change
under the users cursor which seems preferable even if the checkout
takes a couple of seconds.  A spinner or progress bar would be very
beneficial here.