Makeup for fixup commit messages
We unconditionally added a newline (`"\n"`) after the commit subject
which just looks not 100%.  Instead check if we actually have a message
subject already and only then move that subject two(!) lines down so
it becomes part of the message body.

We decided earlier that we *don't* replace old subject lines but only
prepend.  Adding two newlines makes a perfect commit message ready to
commit.