Fix computing the new viewport position `y` value
The old algorithm was not correct if Sublime Text word wrap was set.

Instead of computing the relative `row`, e.g. the cursor is on the 20th
visible row from the top of the viewport, stay in the pixel land and
compute the cursor is "231px" from the top border.  Then after
modifying the view, read the new cursor position (`cy`) in pixel via
`text_to_layout` and set the viewport so that we're still "231px" from
the top (`cy - row_offset`).