Allow restarting conflict resolution of already staged files
Instead of just restarting conflicting files it is desirable to do
this even after staging (and hence "resolving") a file.

To get such a file list - of both resolved and unresolved conflicting
files - we read the proposed `MERGE_MSG` file which actually has this
information in their comment section.  (Is this too naive?)