Fix `LogMixin` not closing the commit info panel
Sublime Text calls `on_highlight` a second time on "done", t.i. with
the same `idx` (or here: `commit`).  Because of that we have a race
between `on_highlight` wanting to show the panel and `on_done`
attempting to close it.

Limit the side-effects of `on_highlight` by applying `lru_cache`.  (We
already solved that in our `show_panel` abstraction.)