Fix: Resilient parsing of the last tag
Fixes #1258

We used `git describe` which correctly gives the latest tag but we're
only interested in the last semver tag.  Just use our `get_local_tags`
instead.

Also be more forgiving when parsing that found tag.  We don't need to
give up here but continue with an input handler as normal.  (The user
can then alter the tag name. If not, she will get the error that a tag
with that name already exists which is probably okay.)