Allow better subclassing of classes using `WithProvideWindow`
Here `super(WithProvideWindow, cloned).run_(...)` is not flexible.
Naturally we want just `super()` to let python handle the mru but
can't because we want to run that on the clone.

Just don't use `super` and call `run_` manually which runs from top to
bottom again.  To restrict the recursion check the presence of
`self.window`.