More robust extraction of issue references
In the "graph" we captured e.g. `Fixes #3` (as a whole) as issue
reference.  Also `aziz/git#4` is recognized as an issue reference.

Add a meta scope `meta.git-savvy.issue-reference` to mark issue
references.  Then check in the code if we actually support the ref
because at the moment we only support refs to the same repo.