Remove `nuke_cursors` as a direct side-effect of drawing
`nuke_cursors` here had the effect of placing the cursor at BOF (`0`)
position.  This is almost never what we want.

The branches dashboard for example then implements `reset_cursor` to
do something different, namely to place the cursor at the active branch.
(This is btw only used after the checkout commands.)

`EditView` used this after the first (and only?) render to put the
cursor at BOF.  This should be done by `replace_view_content` for free.