Run after first render callback immediately
This did not work previously in that the callback run before meaningful
content was rendered, so I had this artificial "async" delay
here (`enqueue_on_worker`).  But it does now and feels so much faster,
well because it is instant now.

Note for testing: we use `on_new_dashboard` to move the cursor to a
meaningful position for new views.