Remove "push all tags" feature
The "push --all" is a non-feature because it could push temporary,
unannotated tags to the remote.

In the tags dashboard a user can select multiple tags to speed thing
up.  Otherwise, a user might have `follow-tags`[1] turned on when
pushing which makes the manual push here obsolete anyway.

[1] "follow-tags" does only push annotated tags so the main confusion
cannot occur then.