Fix `previous_commit`
If `follow` is on and the `current_commit` is the one where the rename
happens, skip 1 actually skips the rename and we will thus not get the
correct result.

Hence, always take 2 commits and skip the first here in python land.

This is in effect only in the blame view because `follow` is not used
anywhere else.