Fix ill filename following in blame view
Following potential filename changes is completely broken in the blame
view.  First, `filename_at_commit` without `follow` on makes no sense
as we only get the input filename if we don't follow actually renames.

Then the parsing was wrong.  Since we go from `<current_commit>..`
this exact commit is not part of the output thus:

```
	if lines[i].split(" ")[0][:commit_len] == commit_hash:
```

will never match.  We also have to take the left side `[1]` of the
the rename because we're interested in the old ("from") name.

We can use `format=` -- t.i. an empty format -- to really only get
"name-status" information.  That makes parsing the output easier.

Note that the comment "git blame does not follow file name changes" is
completely wrong.  `git blame` *always* follows all renames in its
*output*, we just don't use this information (yet).  Of course the
`commit_hash, file_path` tuple as its *input* must be coherent; so we
must track renamings as well.