Enable `word_wrap` on the output panel
We go here for the simple switch and let Sublime do.  We only really
want (or "need") this for the first line when we print the git command.

Using the native wrap we KISS and allow simple copying of such a line.
(If we would wrap in python *just* the first line, we would multiple
lines which are not as nice to copy afterwards.)