Only `--rebase-merges` if a merge commit is within reach of the rebase
`rebase_merges` adds `label x`, `onto y` commands to the todo list even
if there are no merges within the rebase range (and thus no actual
merge commands in the todo list).

This is a bit ugly as the `(done/todo)` counter is a bit confusing; and
in case the user actually edits the todo list this list is a bit noisier
than usual.  So check if a `--rebase-merges` would be a noop first and
don't set it if so.

Note: `update_refs` does not add anything noisy to the todo list.