Do not set `show_at_center` when navigating
`GsNavigate` has the flag `show_at_center` which was implemented but
unused, t.i. it was `True` for all cases.

Just calling `view.show_at_center` for *every* navigation leads to
permanent scrolling (as soon as the buffer is longer than the viewport).
For `show_at_center=False` we use our self-made `show_region` which
minimizes scrolls and is used in the graph view.

When we minimize scrolling we should feed the implementation with
full regions we want to see.  When we only yield single lines scrolling
can be minimal but of course we want to scroll to for example a hunk
and not a single hunk header line.

This is simple enough implemented in `gs_diff_navigate` where the
"thing" we want to see is the whole region between the lines we selected
before.