Try to show the buffer start for the top-most jump position
As we typically have some header, help text etc at the top of our
views, try to show it when jumping to the first jump position of a
particular view.

This use-case is basically supported already by `show_region` with
the `prefer_end` switch.  We feed in a wide region starting from `0`
but tell the algorithm to show the end of the region if it doesn't fit
completely in the viewport.