Fix not determining a correct syntax at all
Regression since 15ebebf2 (Use new `find_syntax_for_file` function in
ST4).

For ST4 we never start a thread to read all syntaxes so the global
`determine_syntax_thread` stays `None` all the time.  Adjust the
conditional, in fact for ST4 we are always ready to determine a correct
syntax.