Dynamically update help text when unstaging is allowed
We can't have a static help text because `include_unstaged` can change
during the life-time of a commit view.  Therefore, render the help text
on refresh.

We also initially (and synchronously) print the help when creating the
view.  This is important because the syntax relies on the comments
markers of the help text; and the code relies on the comment markers
to read-out the commit message etc.