Fix `distinct_until_state_changed` decorator
The decorator was applied at class creation time and thus the internal
storage `previous_state` was a shared single dictionary for all status
views.

The visible bug was: when closing and then recreating a status view
with the exact same state, the decorator prohibited the initial draw
of the new view.

One could make the comparison more complex, for example by also
comparing `self.view.id()` or so; but the `weakref` implementation is
actually easier and its intended, "correct" usage.

We simply store *additional* information about an object (`self`)
separate from it.