Colorize graph art directly at the cursor
Until now, we always searched for the commit dot on the line of the
cursor.  But often we ask: where does this line here begin or end?

Therefore, allow putting the cursor exactly *on* or into the graph art
and follow that graph from there.  E.g. place the cursor on the right
or left of a `|` symbol.