Remove lock during `just_render`
Since #1622 rendering the template is pure so we can safely remove
locking here.

It is still forbidden to call `update_state` in the render ("section")
functions because it likely yields to a confusing output mixing old and
new data portions.  But since we do not acquire the lock, the program
will not dead-lock in that case anymore.

Aside: we keep the lock in `update_state` because we cannot guarantee
it is called only from one thread sequentially.