Jumping backwards ("up") still compare `region.a` positions
Traditionally we only ever marked one liners so choosing `a` or `b`
wasn't very prominent but it effects basically where we jump "upwards"
when the cursor is within the region.

If for example we mark whole hunks and then jump while being within
the hunk, comparing to `b` (as before) jumped to the previous hunk, and
comparing with `a` (as now) jumps to the beginning of the current hunk.