Stage deleted lines before or after the triangle marker
For deletions only regions, Sublime draws a triangle in the gutter which
naturally sits between two lines.  As such a user cannot be exactly on
such a region.

We needed to special case here but choose to only mark this region when
the cursor was right above the triangle.  This is actually annoying
so we now enable marking such regions from above or below.

It is unclear if the prior - a bit akward? - implementation was actually
required by some means.