Add `Move <current_branch> to <cursor>` command
We already had a "move" action which was a bit awkward.  Restrict the
latter to show up only when HEAD is detached and you basically pull
a branch to your position.

That was the only useful function of the feature anyway:  you checked
out the upstream and then wanted to move the local branch label to it.

Or you tested out a previous commit and then "reset" a branch to that
position.