Fix upwards navigation in the dashboards
Fixes #1712

When comparing `region.a < current_position` we must account for
possible offset we apply automatically.

E.g. if the user is exactly on position `region.a + 4` in one of the
dashboards (4 being the offset there), she is of course *on* the jump
point and wants to go to the line above.

Just to be correct, do the `min` dance with `region.b`.