Fix remote tags filter
For the remote tags we only show tags we don't already show in the local
section.  We used to compare the `sha`s only here but of course there
can be multiple tags on the same commit.

We construct new tuples here just for the comparison because the local
tags also come with datetimes.  A simple `tag in seen` is therefore
not possible.