Fix `blame` not rendering anything
Fixes #1717

`gs_blame_refresh` draws using `gs_new_content_and_regions` but not
using the `regions` part.  That was not implemented since the
optimizations in 6a04d2bb (Draw only the changed parts of a view).

Fix that but also change the caller (`gs_blame_refresh`) to use the
raw `replace_view_content` function which is what all the other
non-dashboard views do.